Lab2/requirements.txt
gostcrypto
pycryptodome